[fix]パネルの消すときの見た目の処理を別のスクリプトに移動
